+++ internal/app/file_replace_other.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package app
+
+import "os"
+
+func replaceFile(source, destination string) error {
+	return os.Rename(source, destination)
+}